Updates to SpinCamera to better support auto exposure
SpinCamera updated to clamp current exposure to the auto exposure min/max range when enabling auto exposure and trigger method was changed to update exposure attribute when running in auto exposure mode so the exposure attribute reflects the camera's exposure prior to triggering. When running CamtrawlAcquisition, this ensures that the proper exposure time is sent to the controller and the strobes are illuminated for the correct amount of time.

THIS IS NOT FULLY TESTED YET.